+++ src/api/tools.ts
@@ -1,3 +1,4 @@
+import { TerminalType } from "@/models/models";
 import request from "@/request";
 export const createNewProject = async (configurations: any): Promise<any> => {
     return request.post("/api/tools/createProject", configurations);
@@ -7,14 +8,18 @@ export const requestAutoComplete = async (configurations: { code: string, pos: n
     return request.post("/api/tools/autoComplete", configurations);
 };
 
-export const createPTY = async (cmd: string | string[]): Promise<{ termID: string }> => {
-    return (await request.post('/api/pty/create', { cmd })).data
+export const createPTY = async (cmd: string | string[], name: string): Promise<TerminalType> => {
+    return (await request.post('/api/pty/create', { cmd, name })).data
 }
 
 export const resizePTY = async (termID: string, rows: number, cols: number): Promise<boolean> => {
     return (await request.post("/api/pty/resize", { termID, rows, cols }))
 }
 
-export const allActivePTYs = async (): Promise<string[]> => {
-    return (await request.get('/api/pty/all')).data.terms
+export const closePTY = async (termID: string): Promise<void> => {
+    return (await request.post("/api/pty/close", { termID }))
+}
+
+export const allActivePTYs = async (): Promise<TerminalType[]> => {
+    return (await request.get('/api/pty/all')).data
 }