+++ src/components/terminal/TerminalTabs.vue
@@ -1,33 +1,73 @@
 <template>
   <el-tabs
     v-model="currentTermID"
-    class="mld-terminal-tabs"
+    :class="{ 'mld-terminal-tabs': true }"
     @tab-click="handleTabClick"
   >
     <el-tab-pane
-      label="User"
-      :name="termID"
-      v-for="termID in termIDs"
-      :key="termID"
+      :name="term.id"
+      v-for="term in terms"
+      :key="term.id"
     >
-      <terminal-view :termID="termID" :ref="termID"></terminal-view>
+      <template #label>
+        <el-popover
+          placement="bottom-start"
+          :width="200"
+          trigger="hover"
+          :show-after="500"
+        >
+          <template #reference>
+            <span class="term-label">
+              <div
+                :class="{
+                  'status-indicator': true,
+                  'term-open': !term.closed,
+                  'term-closed': term.closed,
+                }"
+              ></div>
+              <span style="margin-left: 6px">{{ term.name }}</span>
+            </span>
+          </template>
+          Status: {{ term.closed ? "Stopped" : "Running" }}
+          <icon-button
+            v-if="term.closed"
+            @click="onRestartRequest(term.id)"
+            icon="run"
+            text="Restart"
+          ></icon-button>
+          <icon-button
+            v-else
+            @click="onCloseRequest(term.id)"
+            icon="stop"
+            text="Soft stop"
+          ></icon-button>
+          <!-- <el-button v-if="term.closed">Restart</el-button> -->
+          <!-- <el-button v-else >Stop</el-button> -->
+        </el-popover>
+      </template>
+      <terminal-view :termID="term.id" :ref="term.id"></terminal-view>
     </el-tab-pane>
   </el-tabs>
 </template>
 
 <script lang="ts">
-import { allActivePTYs } from "@/api/tools";
 import TerminalView from "./TerminalView.vue";
+import IconButton from "@/components/basic/IconButton.vue";
+import { allActivePTYs, closePTY } from "@/api/tools";
 import { createPTY } from "@/api/tools";
 import { ElNotification, TabsPaneContext } from "element-plus";
 import { defineComponent } from "vue-demi";
 import { registerOnRunCommandRequest } from "../events/globalevents";
+import { addOnMessageHandler } from "@/api/ws";
+import store from "@/store";
+import { TerminalType } from "@/models/models";
+
 export default defineComponent({
-  components: { TerminalView },
+  components: { TerminalView, IconButton },
   data() {
     return {
       currentTermID: "",
-      termIDs: [] as string[],
+      terms: [] as TerminalType[],
     };
   },
   methods: {
@@ -39,42 +79,81 @@ export default defineComponent({
       console.log("terminal", terminalVue);
       (terminalVue as any).injectCMD(terminal);
     },
+    findIdleTermTab() {
+      return this.terms.find((term) => term.closed);
+    },
+    onCloseRequest(termID: string) {
+      closePTY(termID);
+    },
+    onRestartRequest(termID: string) {
+      const currentTerm = this.terms.find((term) => term.id == termID);
+      if (currentTerm == null) {
+        throw Error("Current term was null");
+      }
+      createPTY(currentTerm.command, currentTerm.name).then(
+        (newPTY: TerminalType) => {
+          currentTerm.closed = newPTY.closed;
+          currentTerm.id = newPTY.id;
+          this.currentTermID = newPTY.id;
+        }
+      );
+    },
   },
   beforeMount() {
-    allActivePTYs().then((ptys: string[]) => {
+    allActivePTYs().then((ptys: TerminalType[]) => {
       console.log("ptys", ptys);
       if (ptys.length == 0) {
-        createPTY("bash")
-          .then((terminal: { termID: string }) => {
-            console.log("created pty:", terminal.termID, terminal);
-            ElNotification.error("PTY connection success!");
-            this.termIDs = [terminal.termID];
+        createPTY("bash", "Terminal")
+          .then((terminal: TerminalType) => {
+            console.log("created pty:", terminal.id, terminal);
+            ElNotification.success("PTY connection success!");
+            this.terms = [terminal];
+            this.currentTermID = terminal.id;
           })
           .catch(() => {
             ElNotification.error("Error when creating PTY!");
           });
       } else {
-        this.termIDs = ptys;
+        this.currentTermID = ptys[0].id;
+        this.terms = ptys;
       }
     });
 
-    registerOnRunCommandRequest((cmd: string) => {
-      console.log(cmd);
-      // return;
-      createPTY(["bash", "-c", cmd])
-        .then((terminal: { termID: string }) => {
-          console.log("created pty:", terminal.termID, terminal);
-          ElNotification.error("PTY connection success!");
-          this.termIDs.push(terminal.termID);
-          this.currentTermID = terminal.termID;
-          // window.setTimeout(() => {
-          //   this.sendCommand(terminal.termID, cmd + "&& exit");
-          // }, 1000);
+    registerOnRunCommandRequest(({ cmd, termName }) => {
+      if (termName == null || termName == "") {
+        ElNotification.error(`Terminal name undefined for command ${cmd}!`);
+        return;
+      }
+      createPTY(
+        ["bash", "-c", `cd ${(store.state as any).controls.cwd} &&` + cmd],
+        termName
+      )
+        .then((terminal: TerminalType) => {
+          console.log("created pty:", terminal.id, terminal);
+          ElNotification.success("PTY connection success!");
+          const idleTerm = this.findIdleTermTab();
+          if (idleTerm == null) {
+            this.terms.push(terminal);
+          } else {
+            idleTerm.id = terminal.id;
+            idleTerm.name = termName;
+            idleTerm.closed = false;
+          }
+          this.currentTermID = terminal.id;
         })
         .catch(() => {
           ElNotification.error("Error when creating PTY!");
         });
     });
+    addOnMessageHandler(
+      "pty-status-change",
+      (msg: { output: string; termID: string }) => {
+        const term = this.terms.find((term) => term.id == msg.termID);
+        if (term != null) {
+          term.closed = true;
+        }
+      }
+    );
   },
 });
 </script>
@@ -82,5 +161,25 @@ export default defineComponent({
 <style scoped>
 .mld-terminal-tabs .el-tabs__header {
   margin: 0 0;
+}
+
+.term-label {
+  display: flex;
+  align-items: center;
+}
+
+.status-indicator {
+  height: 12px;
+  width: 12px;
+  border-radius: 6px;
+  display: inline-block;
+}
+
+.term-open {
+  background-color: #11cb25;
+}
+
+.term-closed {
+  background-color: #ce2a2a;
 }
 </style>