+++ ReleaseNotes.md
@@ -2,9 +2,11 @@
 
 - Reconnection of terminals
 
-## Feature ideas
-- Register events for file type. The shortcut should be customizable. There should be buttons on filetree node, and buttons in the popover should 
-be custimizable.
+## Architecutre Ideas
+- Register events for file type. The shortcut should be customizable. There should be buttons on filetree node, and buttons in the popover should be custimizable.
+
+## Feature Ideas
+- Search key strings inside xlsx files
 
 ## Version 0.3.0
 ### New Features: